backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
python-multipart==0.0.22
instagrapi==2.3.0
openai==2.30.0
pydantic==2.12.5
pyotp==2.9.0
langgraph==1.1.4
langgraph-checkpoint==4.0.1
SQLAlchemy==2.0.48
psycopg[binary]==3.3.3
alembic==1.18.4
cryptography>=43.0.0
PyJWT>=2.8.0
httpx==0.28.1
python-dotenv==1.2.2
